Added error checking for all DB queries, finished changes to INSERT queries

+++ admin/field_processor.php
@@ -2,6 +2,8 @@
 
 include("globals.php");
 
+foreach($_REQUEST as $key => $value) $_REQUEST[$key] = stripslashes($value);
+
 $replace = array("'",'"','&',"\\",':',';','`','[',']');
 
 if ($_POST["action"] == "do_add"){
@@ -16,33 +18,31 @@ if ($_POST["action"] == "do_add"){
 	
 	// Check for duplicate fields
 	
-	$_POST["name"] = stripslashes($_POST["name"]);
 	$_POST["name"] = str_replace($replace,"",$_POST["name"]);
 	$_POST["name"] = str_replace("_"," ",$_POST["name"]);
-	$_POST["name"] = trim(addslashes($_POST["name"]));
+	$_POST["name"] = trim($_POST["name"]);
 	
 	$query = "SELECT `id` FROM `anyInventory_fields` WHERE `name`='".$_POST["name"]."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 	
 	if ($result->numRows() > 0){
 		header("Location: ../error_handler.php?eid=0");
 		exit;
 	}
 	else{
-		$_POST["values"] = stripslashes($_POST["values"]);
-		$_POST["values"] = str_replace("'","",$_POST["values"]);
-		$_POST["values"] = trim(str_replace($replace,"",$_POST["values"]));
+		$_POST["field_values"] = str_replace("'","",$_POST["field_values"]);
+		$_POST["field_values"] = trim(str_replace($replace,"",$_POST["field_values"]));
 		
-		$_POST["default_value"] = stripslashes($_POST["default_value"]);
 		$_POST["default_value"] = str_replace("'","",$_POST["default_value"]);
 		$_POST["default_value"] = trim(str_replace($replace,"",$_POST["default_value"]));
 		
 		if (($_POST["input_type"] == "select") || ($_POST["input_type"] == "radio")){
-			if (($_POST["default_value"] == '') || (stristr($_POST["values"],$_POST["default_value"]) === false)){
+			if (($_POST["default_value"] == '') || (stristr($_POST["field_values"],$_POST["default_value"]) === false)){
 				header("Location: ../error_handler.php?eid=1");
 				exit;
 			}
-			elseif($_POST["values"] == ''){
+			elseif($_POST["field_values"] == ''){
 				header("Location: ../error_handler.php?eid=8");
 				exit;
 			}
@@ -60,26 +60,28 @@ if ($_POST["action"] == "do_add"){
 			$_POST["size"] = 255;
 		}
 		
-		$values = explode(",",$_POST["values"]);
+		$field_values = explode(",",$_POST["field_values"]);
 		
-		if (is_array($values)){
-			foreach($values as $key => $value){
-				$values[$key] = trim($value);
+		if (is_array($field_values)){
+			foreach($field_values as $key => $value){
+				$field_values[$key] = trim($value);
 			}
 			
-			$values = array_unique($values);
+			$field_values = array_unique($field_values);
 		}
 		else{
-			$values = array();
+			$field_values = array();
 		}
 		
-		$values = serialize($values);
+		$field_values = serialize($field_values);
 		$categories = array(0);
 		$categories = serialize($categories);
 		
 		// Get the field order for this field.
 		$query = "SELECT MAX(`importance`) as `biggest` FROM `anyInventory_fields`";
-		$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+		$result = $db->query($query);
+		if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
+		
 		$row = $result->fetchRow();
 		$importance = $row['biggest'] + 1;
 		
@@ -87,13 +89,14 @@ if ($_POST["action"] == "do_add"){
 		$query_data = array("id"=>get_unique_id('anyInventory_fields'),
 							"name"=>$_POST["name"],
 							"input_type"=>$_POST["input_type"],
-							"values"=>$values,
-							"default_value"=>stripslashes($_POST["default_value"]),
-							"size"=>$_POST["size"],
+							"field_values"=>$field_values,
+							"default_value"=>$_POST["default_value"],
+							"size"=>intval($_POST["size"]),
 							"categories"=>$categories,
-							"importance"=>$importance,
+							"importance"=>intval($importance),
 							"highlight"=>intval(($_POST["highlight"] == "yes")));
-		$db->autoExecute('anyInventory_fields',$query_data,DB_AUTOQUERY_INSERT);
+		$result = $db->autoExecute('anyInventory_fields',$query_data,DB_AUTOQUERY_INSERT);
+		if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 		
 		$field = new field(get_unique_id('anyInventory_fields') - 1);
 		
@@ -107,15 +110,18 @@ if ($_POST["action"] == "do_add"){
 }
 elseif($_GET["action"] == "do_add_divider"){
 	$query = "SELECT MAX(`importance`) as `biggest` FROM `anyInventory_fields`";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
-	$row = $db->fetchRow();
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
+	
+	$row = $result->fetchRow();
 	$importance = $row['biggest'] + 1;
 	
 	$query_data = array("id"=>get_unique_id('anyInventory_fields'),
 						"name"=>'divider',
 						"input_type"=>'divider',
 						"importance"=>$importance);
-	$db->autoExecute('anyInventory_fields',$query_data,DB_AUTOQUERY_INSERT);
+	$result = $db->autoExecute('anyInventory_fields',$query_data,DB_AUTOQUERY_INSERT);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 }
 elseif($_POST["action"] == "do_edit"){
 	if (is_array($_POST["add_to"])){
@@ -128,7 +134,8 @@ elseif($_POST["action"] == "do_edit"){
 	}
 	
 	$query = "SELECT `id` FROM `anyInventory_fields` WHERE `name`='".$_POST["name"]."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 	
 	$numrows = $result->numRows();
 	$row = $result->fetchRow();
@@ -138,25 +145,22 @@ elseif($_POST["action"] == "do_edit"){
 		exit;
 	}
 	
-	$_POST["values"] = stripslashes($_POST["values"]);
-	$_POST["values"] = str_replace("'","",$_POST["values"]);
-	$_POST["values"] = trim(str_replace($replace,"",$_POST["values"]));
+	$_POST["field_values"] = str_replace("'","",$_POST["field_values"]);
+	$_POST["field_values"] = trim(str_replace($replace,"",$_POST["field_values"]));
 	
-	$_POST["default_value"] = stripslashes($_POST["default_value"]);
 	$_POST["default_value"] = str_replace("'","",$_POST["default_value"]);
 	$_POST["default_value"] = trim(str_replace($replace,"",$_POST["default_value"]));
 	
-	$_POST["name"] = stripslashes($_POST["name"]);
 	$_POST["name"] = str_replace($replace,"",$_POST["name"]);
 	$_POST["name"] = str_replace("_"," ",$_POST["name"]);	
-	$_POST["name"] = trim(addslashes($_POST["name"]));
+	$_POST["name"] = trim($_POST["name"]);
 	
 	if (($_POST["input_type"] == "select") || ($_POST["input_type"] == "radio")){
-		if (($_POST["default_value"] == '') || (stristr($_POST["values"],$_POST["default_value"]) === false)){
+		if (($_POST["default_value"] == '') || (stristr($_POST["field_values"],$_POST["default_value"]) === false)){
 			header("Location: ../error_handler.php?eid=1");
 			exit;
 		}
-		elseif($_POST["values"] == ''){
+		elseif($_POST["field_values"] == ''){
 			header("Location: ../error_handler.php?eid=8");
 			exit;
 		}
@@ -181,31 +185,32 @@ elseif($_POST["action"] == "do_edit"){
 		$_POST["size"] = '';
 	}
 	
-	$values = explode(",",$_POST["values"]);
+	$field_values = explode(",",$_POST["field_values"]);
 	
-	if (is_array($values)){
-		foreach($values as $key => $value){
-			$values[$key] = trim($value);
+	if (is_array($field_values)){
+		foreach($field_values as $key => $value){
+			$field_values[$key] = trim($value);
 		}
 		
-		$values = array_unique($values);
+		$field_values = array_unique($field_values);
 	}
 	else{
-		$values = array();
+		$field_values = array();
 	}
 	
-	$values = serialize($values);
+	$field_values = serialize($field_values);
 	
 	// Change the field.
 	$query = "UPDATE `anyInventory_fields` SET
 				`name`='".$_POST["name"]."',
 				`input_type`='".$_POST["input_type"]."',
-				`values`='".$values."',
+				`field_values`='".$field_values."',
 				`default_value`='".$_POST["default_value"]."',
-				`size`='".$_POST["size"]."',
+				`size`='".intval($_POST["size"])."',
 				`highlight`='".((int) (($_POST["highlight"] == "yes") / 1))."'
 				WHERE `id`='".$_POST["id"]."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 	
 	// Make an object from the new field.
 	$new_field = new field($_POST["id"]);
@@ -238,11 +243,14 @@ elseif($_REQUEST["action"] == "do_delete"){
 		
 		if ($field->input_type == 'file'){
 			$query = "SELECT `value` FROM `anyInventory_fields` WHERE `field_id`='".$field->id."' GROUP BY `value`";
-			$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+			$result = $db->query($query);
+		 	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 			
 			while ($row = $result->fetchRow()){
 				$newquery = "SELECT * FROM `anyInventory_files` WHERE `id`='".$row["value"]."'";
-				$newresult = $db->query($newquery) or die($db->error() . '<br /><br />'. $newquery);
+				$newresult = $db->query($newquery);
+				if (DB::isError($newresult)) die($newresult->getMessage().': line '.__LINE__.'<br /><br />'.$newresult->userinfo);
+				
 				$newrow = $newresult->fetchRow();
 				
 				$file = new file_object($newrow["id"]);
@@ -252,23 +260,27 @@ elseif($_REQUEST["action"] == "do_delete"){
 				}
 				
 				$newestquery = "DELETE FROM `anyInventory_files` WHERE `id`='".$file->id."'";
-				$db->query($newestquery) or die($db->error() . '<br /><br />'. $newestquery);
+				$newestresult = $db->query($newestquery);
+				if (DB::isError($newestresult)) die($newestresult->getMessage().': line '.__LINE__);
 			}
 		}
 		
 		// Change the importance of the fields below it.
 		$query = "UPDATE `anyInventory_fields` SET `importance`=(`importance` - 1) WHERE `importance` >= '".$field->importance."'";
-		$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+		$result = $db->query($query);
+		if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 		
 		if ($field->input_type != 'divider'){
 			// Remove the field from the items table
 			$query = "DELETE FROM `anyInventory_values` WHERE `field_id`='".$field->id."'";
-			$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+			$result = $db->query($query);
+			if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 		}
 		
 		// Delete the field 
 		$query = "DELETE FROM `anyInventory_fields` WHERE `id`='".$field->id."'";
-		$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+		$result = $db->query($query);
+		if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 	}
 }
 elseif($_GET["action"] == "moveup"){
@@ -279,10 +291,12 @@ elseif($_GET["action"] == "moveup"){
 	
 	// Move a field up
 	$query = "UPDATE `anyInventory_fields` SET `importance`=".$_GET["i"]." WHERE `importance`='".($_GET["i"] - 1)."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 	
 	$query = "UPDATE `anyInventory_fields` SET `importance`=".($_GET["i"] - 1)." WHERE `id`='".$_GET["id"]."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 }
 elseif($_GET["action"] == "movedown"){
 	if (!$admin_user->can_admin_field($_GET["id"])){
@@ -292,10 +306,12 @@ elseif($_GET["action"] == "movedown"){
 	
 	// Move a field down
 	$query = "UPDATE `anyInventory_fields` SET `importance`=".$_GET["i"]." WHERE `importance`='".($_GET["i"] + 1)."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 	
 	$query = "UPDATE `anyInventory_fields` SET `importance`=".($_GET["i"] + 1)." WHERE `id`='".$_GET["id"]."'";
-	$result = $db->query($query) or die($db->error() . '<br /><br />'. $query);
+	$result = $db->query($query);
+	if (DB::isError($result)) die($result->getMessage().': line '.__LINE__.'<br /><br />'.$result->userinfo);
 }
 
 header("Location: fields.php");
